# Size tracking bitvector