A couple of notes:
- I didn't fully understand why the original representation used only 62 bits. I changed it to use 63. Maybe there was some confusion about the hamming distance/count bits code? But I believe that it also works for negative values.
- I changed slightly the behavior or `of_list`, `length` and `resize`.
- In general, the more that I worked on this the more that I'm puzzled about the actual usefulness of bitvectors that resize. I understand that it would be convenient, but for most use cases, the user is tracking a known set of states. In these cases I would imagine it an error for things to somewhat silently change size. It also leads to weird cases where `let i = 10000000 in get t i`  will succeed but then `set t i` might allocate a big chunk of memory. There could would seem more straight forward with fixed sizes and I'm tempted to use such a subset in my internal project.

Here are some micro benchmarks:

Creation

| Name     | Runs @ Samples | Time R^2 | Time/Run |            95ci | Cycls/Run | mWd/Run |
|---------:|---------------:|---------:|---------:|----------------:|----------:|--------:|
| Old:10   |    69757 @ 815 |     1.00 |   1.43us | -0.01us +0.01us |    4.28kc | 146.00w |
| Old:100  |     6517 @ 576 |     1.00 |  15.00us | -0.06us +0.07us |   45.01kc | 233.00w |
| Old:1000 |      582 @ 325 |     1.00 | 155.70us | -0.85us +1.13us |  467.09kc | 246.00w |
| New:10   |    74045 @ 821 |     1.00 |   1.35us | -0.01us +0.01us |    4.04kc |  74.00w |
| New:100  |     6582 @ 577 |     1.00 |  14.77us | -0.07us +0.08us |   44.32kc |  74.00w |
| New:1000 |      587 @ 326 |     1.00 | 155.06us | -0.73us +0.79us |  465.15kc |  74.00w |

Union

| Name     | Runs @ Samples | Time R^2 |     Time/Run |                95ci | Cycls/Run |    mWd/Run |
|---------:|---------------:|---------:|-------------:|--------------------:|----------:|-----------:|
| Old:10   |    83428 @ 833 |     0.98 |   1_202.80ns |   -30.65ns +51.24ns |    3.61kc |    619.00w |
| Old:100  |     7410 @ 589 |     0.98 |  13_180.18ns | -209.43ns +341.70ns |   39.54kc |  6_739.00w |
| Old:1000 |      686 @ 343 |     1.00 | 133_154.77ns | -405.34ns +415.45ns |  399.44kc | 67_939.00w |
| New:10   |   121743 @ 871 |     0.99 |     822.18ns |   -14.10ns +23.16ns |    2.47kc |    619.00w |
| New:100  |    10897 @ 628 |     0.99 |   9_041.62ns |  -88.02ns +106.60ns |   27.12kc |  6_739.00w |
| New:1000 |      987 @ 382 |     0.99 |  92_881.11ns | -639.35ns +855.09ns |  278.63kc | 67_939.00w |

Difference

| Name     | Runs @ Samples | Time R^2 |    Time/Run |                95ci |   Cycls/Run |     mWd/Run |
|---------:|---------------:|---------:|------------:|--------------------:|------------:|------------:|
| Old:10   |      324 @ 258 |     1.00 |    285.11us |     -1.11us +1.22us |    855.30kc |   1_276.01w |
| Old:100  |        81 @ 81 |     1.00 |  3_146.65us |   -24.91us +31.65us |  9_439.57kc |  13_966.00w |
| Old:1000 |        26 @ 26 |     1.00 | 31_455.14us | -151.50us +148.36us | 94_362.01kc | 140_866.00w |
| New:10   |    86814 @ 837 |     0.97 |      1.12us |     -0.01us +0.02us |      3.37kc |     619.00w |
| New:100  |     8020 @ 597 |     0.99 |     12.25us |     -0.21us +0.30us |     36.74kc |   6_739.00w |
| New:1000 |      746 @ 352 |     1.00 |    122.61us |     -0.60us +0.62us |    367.80kc |  67_939.00w |


Happy to discuss and document and test further.