@c-cube 2017-04-19
I was not totally sure of the semantics of `lsl` w.r.t the sign bit, so I played it safe and removed it altogether, but you're right, I don't think I use anything signed anyway. So 63 bits should be safe.