@c-cube 2017-04-19
Also: there is a benchmark suite in containers already, it would be better to use it if you can :-)
It's possible to copy the current implem into it as a "reference", I'm not afraid of code duplication in benchmarks.