@rleonid 2017-04-19
Do you mind if I just grab the changes from #116 without playing `git` merging games?