@rleonid 2017-04-19
Ok, back to you. I'm going to pass on adding to the benchmarks as I don't have the bandwith at the moment.