@rleonid 2017-04-19
Ok, added a much needed test.